- Extracting to a different file those methods that support the testing process
